C++: add asserts that we are in the main thread
 - In the timer functions
 - in the generated getter/setters and callback invocation
